// fitepic-web-app/src/app/core/api/generated/models/workout-score-type.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export type WorkoutScoreType = 'None' | 'TimeToComplete' | 'TimeCapPlusReps' | 'TotalReps' | 'RepsPerRound' | 'HeaviestLoad' | 'TotalLoad' | 'TimeAndLoad' | 'RepsAndLoad' | 'PointsSystem' | 'RoundsAndReps' | 'Meters' | 'Calories' | 'Miles' | 'Feet' | 'TieBreakTime' | 'CustomNumeric';